Fixed custom spells
Adding custom spells actually works now. I ironed out the kinks, added a
"refresh Spell_List" method to the class. It basically reinitializes it.
I was worried that it would be a huge burden on the program, but it only
takes 0.0015 seconds to refresh and is only reading/writing ~400 KB of
information. It's not like I'm burning out the drive or anything. I'm in
that unfortunate state where I've been working on this code for 3 hours
now and mostly forget what changes I've made beyond that.

Here's a story to remember: I accidentally hid a bug by using good
coding practice. Way way back when I started this project and made the
SortedSpellList class, I added a check to the list where it does not add
a spell if it already exists in the list. In my use case, a spell list
shouldn't have repeats. So when there's this big pesky bug I'm trying to
fix, I can't understand why the fake custom spell isn't going away. Well
it turns out that I never cleared the original spell list before adding
on the updated valeus. So I had shit tons of duplicate spells stored in
the Spell_List class. But since SortedSpellList doesn't allow duplicate
entries, I wasn't able to see this duplicate adding. All I saw was that
a custom spell wasn't getting deleted. It was only because I remembered
a trait of the spell list I coded several months ago that I was able to
find and fix the bug.